Fix to the Generalized Hyperbolic estimation
Fix to the Generalized Hyperbolic Distribution estimation code as well as the scaled besselK function in the presence of fixed lambda. The fix somewhat negates the speedup from the previous version when lambda is fixed, but ensures correctness. To improve the speed, the bessel_k_10 function from adcomp/TMB/inst/include/atomic_math.hpp need to be adjusted to allow scaling and the reverse mode adjusted in the presence of this scaling [to be done/roadmap].